# Optimize io (de)serialization routines