--- a/pkg/consensus/change_view.go
+++ b/pkg/consensus/change_view.go
@@ -15,7 +15,7 @@ type changeView struct {
 var _ payload.ChangeView = (*changeView)(nil)
 
 // EncodeBinary implements io.Serializable interface.
-func (c *changeView) EncodeBinary(w *io.BinWriter) {
+func (c *changeView) EncodeBinary(w io.BinaryWriter) {
 	w.WriteU64LE(c.timestamp)
 	w.WriteB(byte(c.reason))
 }