Hiding `BinWriter` behind an interface allows us to use specialized implementations for buffer writer, counter writer etc.

`BinReader` can be impoved similarly, will do it later.