@fyrchik 2021-08-05
With `WriteArray` removal there are bigger relative gains:
```
name           old time/op    new time/op    delta
EncodeNew-8       614ns ± 6%     510ns ± 1%  -16.98%  (p=0.000 n=10+8)
EncodeReset-8     183ns ± 4%      87ns ± 3%  -52.52%  (p=0.000 n=10+10)
```

Another thing is new reader:
- calculate tx, block and extensible payload hash without re-serizalizing (see last commit, the interface can be improved)
- `ReadBytes` is a simple copy now (instead of intermediate slice allocation)

I think we can try to implement `Grow` and `VarSize` optimizations on top of existing `io` structures and wait with this massive refactoring until something else comes to mind.