@roman-khimov 2022-01-11
I think this one is severely outdated by now. Some recent optimizations also made it less relevant, so I'm closing it.